+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - Unreleased
+### Added
+- Handle array values when decoding simple events
+
+### Fixed
+- Don't treat structs as object aggregations when decoding simple events
+
 ## [1.0.0] - 2021-06-28
 
 ## [1.0.0-rc.0] - 2021-05-05